+++ src/libs/gl/draw.c
@@ -56,7 +56,7 @@ void GL_context_fill(const GL_Context_t *context, GL_Point_t seed, GL_Pixel_t in
     const GL_Surface_t *surface = context->surface;
     const GL_State_t *state = &context->state.current;
     const GL_Quad_t *clipping_region = &state->clipping_region;
-    const uint16_t *state_map = state->palette_state.map;
+    const uint8_t *state_map = state->palette_state.map;
 
     if (seed.x < clipping_region->x0 || seed.x >= clipping_region->x1
         || seed.y < clipping_region->y0 || seed.y >= clipping_region->y1) {
@@ -69,7 +69,7 @@ void GL_context_fill(const GL_Context_t *context, GL_Point_t seed, GL_Pixel_t in
 
     const size_t dskip = dwidth;
 
-    uint16_t mapped = state_map[index];
+    uint8_t mapped = state_map[index];
     if (transparency && GL_PALETTE_IS_TRANSPARENT(mapped)) {
         return;
     }
@@ -133,7 +133,7 @@ void GL_context_scan(const GL_Context_t *context, GL_Rectangle_t area, const GL_
     const GL_Surface_t *surface = context->surface;
     const GL_State_t *state = &context->state.current;
     const GL_Quad_t *clipping_region = &state->clipping_region;
-    const uint16_t *state_map = state->palette_state.map;
+    const uint8_t *state_map = state->palette_state.map;
 
     GL_Quad_t drawing_region = (GL_Quad_t){
             .x0 = area.x,
@@ -173,7 +173,7 @@ void GL_context_scan(const GL_Context_t *context, GL_Rectangle_t area, const GL_
     for (int i = height; i; --i) {
         int x = drawing_region.x0; // TODO: optimize?
         for (int j = width; j; --j) {
-            uint16_t mapped = state_map[callback(user_data, (GL_Point_t){ .x = x, .y = y }, *dptr)];
+            uint8_t mapped = state_map[callback(user_data, (GL_Point_t){ .x = x, .y = y }, *dptr)];
             if (GL_PALETTE_IS_TRANSPARENT(mapped)) {
                 ++dptr;
             } else {
@@ -191,7 +191,7 @@ void GL_context_process(const GL_Context_t *context, GL_Point_t position, const 
     const GL_Surface_t *surface = context->surface;
     const GL_State_t *state = &context->state.current;
     const GL_Quad_t *clipping_region = &state->clipping_region;
-    const uint16_t *state_map = state->palette_state.map;
+    const uint8_t *state_map = state->palette_state.map;
 
     int skip_x = area.x; // Offset into the (source) surface/texture, updated during clipping.
     int skip_y = area.y;
@@ -243,7 +243,7 @@ void GL_context_process(const GL_Context_t *context, GL_Point_t position, const 
             const GL_Pixel_t from = *dptr;
             const GL_Pixel_t to = *(sptr++);
 
-            uint16_t mapped = state_map[callback(user_data, (GL_Point_t){ .x = x, .y = y }, from, to)];
+            uint8_t mapped = state_map[callback(user_data, (GL_Point_t){ .x = x, .y = y }, from, to)];
             if (GL_PALETTE_IS_TRANSPARENT(mapped)) {
                 ++dptr;
             } else {
@@ -368,7 +368,7 @@ void GL_context_stencil(const GL_Context_t *context, GL_Point_t position, const 
     const GL_Surface_t *surface = context->surface;
     const GL_State_t *state = &context->state.current;
     const GL_Quad_t *clipping_region = &state->clipping_region;
-    const uint16_t *state_map = state->palette_state.map; // TODO: should `GL_surface_copy()` and `GL_surface_mask()` skip shifting and transparency?
+    const uint8_t *state_map = state->palette_state.map; // TODO: should `GL_surface_copy()` and `GL_surface_mask()` skip shifting and transparency?
     const GL_Pixel_Comparator_t should_write = _pixel_comparators[comparator];
 
 #if defined(TOFU_CORE_DEFENSIVE_CHECKS)
@@ -432,7 +432,7 @@ void GL_context_stencil(const GL_Context_t *context, GL_Point_t position, const 
 #endif
             const GL_Pixel_t value = *(mptr++);
 
-            uint16_t mapped = state_map[*(sptr++)];
+            uint8_t mapped = state_map[*(sptr++)];
             if (GL_PALETTE_IS_TRANSPARENT(mapped)
                 || !should_write(value, threshold)) {
                 ++dptr;
@@ -522,7 +522,7 @@ void GL_context_blend(const GL_Context_t *context, GL_Point_t position, const GL
     const GL_Surface_t *surface = context->surface;
     const GL_State_t *state = &context->state.current;
     const GL_Quad_t *clipping_region = &state->clipping_region;
-    const uint16_t *state_map = state->palette_state.map;
+    const uint8_t *state_map = state->palette_state.map;
     const GL_Pixel_Function_t blend = _pixel_functions[function];
 
     int skip_x = area.x; // Offset into the (source) surface/texture, update during clipping.
@@ -573,7 +573,7 @@ void GL_context_blend(const GL_Context_t *context, GL_Point_t position, const GL
 #if defined(TOFU_GRAPHICS_DEBUG_ENABLED)
             _pixel(surface, drawing_region.x0 + width - j, drawing_region.y0 + height - i, i + j);
 #endif
-            uint16_t mapped = state_map[blend(*dptr, *(sptr++))];
+            uint8_t mapped = state_map[blend(*dptr, *(sptr++))];
             if (GL_PALETTE_IS_TRANSPARENT(mapped)) {
                 ++dptr;
             } else {
